file addition

+++ test/dataloader/trainbis.py
@@ -94,7 +94,7 @@ metrics = evaluation(n_classes=n_classes, lr=lrs[0], modelstr="SegNet",
 
 weights = WeightComputationMedian(labels_path=label_path, n_classes=n_classes)
 
-weights = torch.from_numpy(weights)
+weights = torch.from_numpy(weights).float()
 
 criterion = nn.CrossEntropyLoss(weight=weights, reduce=True,
                                 size_average=True).cuda()
